Ticket #67 "segment visibility system"
This includes an implementation of segment garbage collection that is
purely SQL based.  Garbage collection related information is consumed by
processes beyond just garbage collection, such as web servers
implementing API features such as listing available keys and available
versions.

This also includes a way to move the implementation of listing keys on a
collection with a key marker and limit, and listing versions with a key
marker, version marker, and limit, entirely into SQL.